requirements.txt
asgiref==3.5.2
distlib==0.3.4
Django==4.0.6
django-filter==22.1
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
filelock==3.7.1
importlib-metadata==4.12.0
Markdown==3.4.1
Pillow==9.2.0
platformdirs==2.5.2
PyJWT==2.4.0
pytz==2022.1
six==1.16.0
sqlparse==0.4.2
virtualenv==20.14.1
zipp==3.8.1